+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add module mix-in functions to `RbObject`.
 * Bind Ruby objects and methods directly to Swift objects and methods.
 * Add throwing conversion as alternative to optional initializer.
+* Add `RbMethod.callSuper()` to call superclass method.
 
 ##### Bug Fixes